// client/src/utils/types/legacyBackup.ts
export interface LegacyBackup {
  Pos_ID: string | null;
  Cor: string;
  Vias: string | null;
  CODIVMAC: string;
  ConnType: string | null;
  Designa__o: string | null;
  Fabricante: string | null;
  Refabricante: string | null;
  Registo: string | null;
  OBS: string | null;
  LastChangeBy: string | null;
  LastUpdateDate: string | null;
  ESTADO: number;
  ClipColor: string | null;
  CapotAngle: string | null;
  InternalDiameter: number | null;
  ExternalDiameter: number | null;
  Thickness: number | null;
}